- update weather's datas
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6677 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: b6233be6a77744ed2c8b804d6044530b9840b4c0
Former-commit-id: cf22c3b8245e9331ba82bb4b08ebef674997682d

+++ src/drivers/shadow/src/CarModel.h
@@ -57,7 +57,7 @@ public:
 
     double  BRAKESCALE;     // Scaling of Brake
     double  BRAKEFORCE;     // Brake force max
-	float	BRAKELIMIT;		// BrakeLimit
+    double	BRAKELIMIT;		// BrakeLimit
 
     double	CA;             // aerodynamic downforce constant -- total.
     double	CA_FW;          // aerodynamic downforce constant -- front wing.
